tweet-generator/requirements.txt
google-cloud-pubsub==0.33.0
Faker==0.8.12
ratelimit==1.4.1